feat: add main config section with header behind feature flag (#14702)
