+++ frontend/packages/ux-editor/src/components/Properties/PropertiesHeader/ComponentMainConfig.test.tsx
@@ -1,22 +1,19 @@
 import React from 'react';
-import { screen } from '@testing-library/react';
-import { renderWithProviders } from '../../../testing/mocks';
-import { ComponentMainConfig } from './ComponentMainConfig';
-import type { FormItem } from '../../../types/FormItem';
+import type { FormItem } from '@altinn/ux-editor/types/FormItem';
 import { ComponentType } from 'app-shared/types/ComponentType';
-import userEvent from '@testing-library/user-event';
+import { screen } from '@testing-library/react';
+import { ComponentMainConfig } from './ComponentMainConfig';
+import { component1Mock } from '@altinn/ux-editor/testing/layoutMock';
+import { addFeatureFlagToLocalStorage, FeatureFlag } from 'app-shared/utils/featureToggleUtils';
+import { textMock } from '@studio/testing/mocks/i18nMock';
+import { typedLocalStorage } from '@studio/pure-functions';
+import { renderWithProviders } from '../../../testing/mocks';
 import { createQueryClientMock } from 'app-shared/mocks/queryClientMock';
 import { QueryKey } from 'app-shared/types/QueryKey';
 import { app, org } from '@studio/testing/testids';
-import {
-  layoutSet1NameMock,
-  layoutSet2NameMock,
-  layoutSetsExtendedMock,
-  layoutSetsMock,
-} from '../../../testing/layoutSetsMock';
-import { layout1NameMock, layoutMock } from '../../../testing/layoutMock';
+import { layoutSetsExtendedMock } from '@altinn/ux-editor/testing/layoutSetsMock';
 
-const summary2Component: FormItem = {
+const summary2ComponentMock: FormItem = {
   id: '0',
   type: ComponentType.Summary2,
   itemType: 'COMPONENT',
@@ -24,75 +21,34 @@ const summary2Component: FormItem = {
 };
 
 describe('ComponentMainConfig', () => {
-  describe('Summary2', () => {
-    beforeEach(() => {
-      jest.clearAllMocks();
-    });
+  afterEach(() => {
+    jest.clearAllMocks();
+    typedLocalStorage.removeItem('featureFlags');
+  });
 
-    it('should render summary2 config', async () => {
-      const user = userEvent.setup();
-      render(summary2Component);
-      expect(summary2AccordionButton()).toBeInTheDocument();
-      await user.click(summary2AccordionButton());
-      expect(summary2AddOverrideButton()).toBeInTheDocument();
-    });
+  it('should render summary2 config when the component type matches', async () => {
+    renderComponentMainConfig(summary2ComponentMock);
 
-    it('should display overrides', async () => {
-      const user = userEvent.setup();
-      const summary2ComponentWithOverrides = {
-        ...summary2Component,
-        overrides: [{ componentId: '0' }],
-      };
-      render(summary2ComponentWithOverrides);
-      await user.click(summary2AccordionButton());
-      expect(summary2CollapsedButton(1)).toBeInTheDocument();
-    });
+    const targetHeader = screen.getByText(textMock('ux_editor.component_properties.target'));
+    expect(targetHeader).toBeInTheDocument();
+  });
 
-    it('should call handleComponentChange when adding overrides', async () => {
-      const user = userEvent.setup();
-      render(summary2Component);
-      await user.click(summary2AccordionButton());
-      await user.click(summary2AddOverrideButton());
-      expect(handleComponentChange).toHaveBeenCalledTimes(1);
-    });
+  it('should render header config when feature flag is set, but the type does not match', async () => {
+    addFeatureFlagToLocalStorage(FeatureFlag.MainConfig);
+    renderComponentMainConfig(component1Mock);
 
-    it('should call handleComponentChange when changing target', async () => {
-      const user = userEvent.setup();
-      render(summary2Component);
-      await user.selectOptions(summary2TargetLayoutSet(), layoutSet2NameMock);
-      expect(handleComponentChange).toHaveBeenCalledTimes(1);
-    });
+    const sectionHeader = textMock('ux_editor.component_properties.main_configuration');
+    const headerMainConfig = screen.getByText(sectionHeader);
+    expect(headerMainConfig).toBeInTheDocument();
   });
 });
 
-const summary2AccordionButton = () =>
-  screen.getByRole('button', { name: /ux_editor.component_properties.summary.override.title/ });
-const summary2AddOverrideButton = () =>
-  screen.getByRole('button', { name: /ux_editor.component_properties.summary.add_override/ });
-const summary2CollapsedButton = (n: number) =>
-  screen.getByRole('button', {
-    name: new RegExp(`ux_editor.component_properties.summary.overrides.nth.*:${n}}`),
-  });
-
-const summary2TargetLayoutSet = () =>
-  screen.getByRole('combobox', { name: /ux_editor.component_properties.target_layoutSet_id/ });
-
-const handleComponentChange = jest.fn();
-const render = (component: FormItem) => {
+const renderComponentMainConfig = (component: FormItem) => {
+  const handleComponentChange = jest.fn();
   const queryClient = createQueryClientMock();
-  queryClient.setQueryData([QueryKey.FormLayouts, org, app, layoutSet1NameMock], {
-    [layout1NameMock]: layoutMock,
-  });
-  queryClient.setQueryData([QueryKey.LayoutSets, org, app], layoutSetsMock);
   queryClient.setQueryData([QueryKey.LayoutSetsExtended, org, app], layoutSetsExtendedMock);
-  renderWithProviders(
+  return renderWithProviders(
     <ComponentMainConfig component={component} handleComponentChange={handleComponentChange} />,
-    {
-      queryClient,
-      appContextProps: {
-        selectedFormLayoutSetName: layoutSet1NameMock,
-        selectedFormLayoutName: layout1NameMock,
-      },
-    },
+    { queryClient },
   );
 };
