+++ frontend/packages/ux-editor/src/components/Properties/PropertiesHeader/HeaderMainConfig.module.css
